Apply old uncommitted changes — still needs final fixes
