+++ app/packages.egg-info/PKG-INFO
@@ -1,0 +1,3 @@
+Metadata-Version: 2.1
+Name: packages
+Version: 1.0
