+++ app/packages.egg-info/SOURCES.txt
@@ -1,0 +1,5 @@
+setup.py
+packages.egg-info/PKG-INFO
+packages.egg-info/SOURCES.txt
+packages.egg-info/dependency_links.txt
+packages.egg-info/top_level.txt
